requirements.txt
aiogram==2.25.2
flask
requests
python-dotenv
aiohttp==3.8.6
